+++ CHANGES.rst
@@ -8,6 +8,7 @@ Version 0.2
 (released on ???)
 
 - Dropped support for Python versions below 2.7.
+- Switched DOT library from 'pydot' to 'graphviz'.
 - Don't require `ez_setup` anymore.
 - Moved package docstring into stand-alone README file.
 - Moved requirement specifications from setup script into pip-style text files.